+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.46
+
+* added kwargs to ThreadCollection.from_class